requirements.txt
Flask>=2.0.0
pandas>=1.3.0
requests>=2.25.0
beautifulsoup4>=4.9.3
openpyxl>=3.0.7
PyMuPDF>=1.18.0  # para o pacote fitz
PyPDF2>=3.0.0
python-docx>=0.8.11
Werkzeug>=2.0.0
